% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{clock_speed}
\alias{clock_speed}
\title{CPU clock speeds}
\format{A two column data frame}
\source{
https://csgillespie.wordpress.com/2011/01/25/cpu-and-gpu-trends-over-time/
}
\description{
The year and clock speed of CPUs. The clock speed is measured in million of instructions
per second. This data set was collected from forums, wikipedia and blog posts.
}

