% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{die}
\alias{die}
\title{MIPS per die}
\format{A three column data frame}
\source{
https://csgillespie.wordpress.com/2011/01/25/cpu-and-gpu-trends-over-time/
}
\description{
The year, clock speed and brief description of CPUs. 
The clock speed is measured in million of instructions
per second per die. This data set was taken from
http://en.wikipedia.org/wiki/List_of_Intel_microprocessors
Each CPU is the best in it's class
}

