% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effOSil.R
\name{effOSil}
\alias{effOSil}
\title{The Efficient Optimum Silhouette algorithm}
\usage{
effOSil(dx, K, initMethod, variant)
}
\arguments{
\item{dx}{A dist object, which can be computed using the stats::dist() function.}

\item{K}{An integer vector specifying the number of clusters. By default, K = 2:12.}

\item{initMethod}{A character vector specifying initialisation methods. By default,
initMethod = "average"; however, to achieve the best initialisation in terms of the ASW,
various initialisation methods should be used (e.g., initMethod = c("single", "average", "complete", "pam")).
See ?Init for more details.}

\item{variant}{An algorithmic variant. Options include "efficient" and "original". By default, variant = "efficient", indicating that effOSil is used.
If variant = "original", the original, computationally expensive OSil algorithm is used.}
}
\value{
\describe{
\item{best_clustering}{The clustering achieving the highest ASW value.}
\item{best_asw}{The highest ASW value.}
\item{k}{The estimated number of clusters.}
\item{clusterings}{The effOSil clusterings for all k in K.}
\item{asw}{The ASW values associated with the clusterings.}
\item{nIter}{The numbers of iterations needed for convergence.}
}
}
\description{
This function implements the Efficient Optimum Silhouette (effOSil) algorithm.
}
\details{
This function implements the Efficient Optimum Silhouette (effOSil) algorithm, an O(N) runtime improvement of
the original, computationally expensive Fast OSil (FOSil) algorithm proposed by Batool & Hennig (2021) where N is
the number of observations. This function also implements the OSil algorithm for comparision purporses.
}
\examples{
x = scale(faithful)
dx = dist(x)
effOSil_clustering = effOSil(dx = dx, K = 2:12)
par(mfrow = c(1,2))
plot(faithful, col = effOSil_clustering$best_clustering, pch = effOSil_clustering$best_clustering)
plot(2:12, effOSil_clustering$asw, type = "l", xlab = "k", ylab = "ASW")
par(mfrow = c(1,1))

}
\references{
Batool, F. and Hennig, C., 2021. Clustering with the average silhouette width. Computational Statistics & Data Analysis, 158, p.107190.
}
\author{
Minh Long Nguyen \email{edelweiss611428@gmail.com}
}
