% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ghg_ems}
\alias{ghg_ems}
\title{Worldwide emissions of greenhouse gases (ghg) by sector, country and year.}
\format{A seven column data frame.}
\description{
Emissions data from the World Resources Institute,
from their website:
\url{http://www.wri.org/resources/data-sets/cait-country-greenhouse-gas-emissions-data}.
}
\examples{
\dontrun{
# After completing the form,
# and downloading the .zip file into the working directory:
unzip("CAIT_Country_GHG_Emissions_-_csv_0.zip")
f <- "CAIT_Country_GHG_Emissions_-_All_Data.xlsx"
ghg_ems <- readxl::read_excel(f, sheet = 4)
head(ghg_ems)
names(ghg_ems) = stringr::word(names(ghg_ems), sep = " |/")
devtools::use_data(ghg_ems, overwrite = TRUE)
}
}

