% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{idata}
\alias{idata}
\title{Global inequality by, country and year.}
\format{A nine column data frame.}
\description{
Estimates of the level of income inequality, using the Gini index,
from World Bank data. See
\url{http://data.worldbank.org/indicator/SI.POV.GINI?locations=AT}
for further information.
}

