% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Init.R
\name{Init}
\alias{Init}
\title{Initialisation methods for the Optimum Silhouette algorithm.}
\usage{
Init(dx, k, initMethod)
}
\arguments{
\item{dx}{dx A dist object, which can be computed using the stats::dist() function.}

\item{k}{An integer specifying the number of clusters.}

\item{initMethod}{A character vector (or string) specifying initialisation methods. Options include any
combination of "pam", "average", "single", "complete", "ward.D", "ward.D2", "mcquitty", "median", and "centroid".
By default, initMethod = "average".}
}
\value{
\describe{
\item{clustering}{An initialised clustering.}
\item{asw}{The ASW associated with the initialised clustering.}
\item{method}{The "best" initialisation method.}
}
}
\description{
This function computes an initialisation for the Optimum Silhouette algorithm.
}
\details{
This function computes an initialisation for the Optimum Silhouette algorithm, but it can be used as
a stand-alone clustering method (i.e., run different clustering algorithms and select the clustering solution maximising the ASW).
}
\examples{
x = faithful
dx = dist(x)
InitClustering = Init(dx, 2, c("pam", "average", "complete"))
plot(x, col = InitClustering$clustering, pch = 4)
print(paste(InitClustering$method, "achieves the highest ASW value"))

}
\references{
Batool, F. and Hennig, C., 2021. Clustering with the average silhouette width. Computational Statistics & Data Analysis, 158, p.107190.
Batool, F., 2019. Initialization methods for optimum average silhouette width clustering. arXiv preprint arXiv:1910.08644.
}
\author{
Minh Long Nguyen \email{edelweiss611428@gmail.com}
}
