% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{lnd_geo_df}
\alias{lnd_geo_df}
\title{Simple geographical/population data of boroughs in London with}
\format{A data frame with 4 variables and 211 rows}
\description{
Generated from the \url{https://github.com/Robinlovelace/Creating-maps-in-R} tutorial using
the simplify function from the mapshaper package and the \code{fortify} function from the ggplot2 package.
}
\examples{
\dontrun{

}
}

