% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monopoly.R
\name{move_square}
\alias{move_square}
\title{Moving square on a Monopoly board}
\usage{
move_square(current)
}
\arguments{
\item{current}{The current square. Number between 1 and 40.
This function calculates where you will land on your next move.}
}
\description{
Moving square on a Monopoly board
}
