% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monopoly.R
\name{simulate_monopoly}
\alias{simulate_monopoly}
\title{Monte Carlo Simulation of Monopoly}
\usage{
simulate_monopoly(no_of_turns)
}
\arguments{
\item{no_of_turns}{How many dice rolls should we take}
}
\description{
An R function for simulating the game of Monopoly
}

