% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snakes_and_ladders.R
\name{snakes_ladders}
\alias{snakes_ladders}
\title{Snakes and Ladders}
\usage{
snakes_ladders(i = 1)
}
\arguments{
\item{i}{A dummy argument needed for the parallel interface.}
}
\value{
Number of rolls needed to win.
}
\description{
A simple game of snakes and ladders.
}

