% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_rcpp.R
\name{test_rcpp}
\alias{test_rcpp}
\title{Rcpp test function}
\usage{
test_rcpp()
}
\value{
Returns \code{TRUE} if everything is OK.
}
\description{
A simple function to test whether Rcpp is installed and working.
}
