% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{transistors}
\alias{transistors}
\title{Transistor counts}
\description{
The number of transistors per CPU. The data was scrapped from
https://en.wikipedia.org/wiki/Transistor_count (see example below)
}
\examples{
\dontrun{
library("rvest")
library("magrittr")

trans_count = read_html("https://en.wikipedia.org/wiki/Transistor_count")
trans_table = trans_count \%>\%
  html_nodes("table") \%>\%
  .[[1]] \%>\% html_table
trans_table[,2] = gsub("\\\\[[0-9]*\\\\]", "", trans_table[,2])
trans_table[,2] = as.numeric(gsub(",|~", "", trans_table[,2]))
trans_table[,3] = as.numeric(gsub("\\\\[[0-9]*\\\\]", "", trans_table[,3]))
colnames(trans_table)[2:3] = c("Count", "Year")
}
}

