% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{wb_ineq}
\alias{wb_ineq}
\alias{wb_ineq_renamed}
\title{Global inequality by, country and year.}
\format{A nine column data frame.}
\description{
Estimates of the level of income inequality, using the Gini index,
from World Bank data. See
\url{http://data.worldbank.org/indicator/SI.POV.GINI?locations=AT}
for further information. \code{wb_ineq_renamed} is a version of the dataset with updated names.
}
\examples{
\dontrun{
head(wb_ineq)
library(dplyr)
wb_ineq_renamed = rename(wb_ineq, code = `Country Code`)
cols_to_change = 5:9 # column ids to change
wb_ineq_renamed = wb_ineq_renamed \%>\% 
  rename(top10 = `SI.DST.10TH.10`, bot10 = `SI.DST.FRST.10`) \%>\% 
  mutate_at(vars(cols_to_change), as.numeric)
}
}
